// src/model/showProductsResponseProductDto.ts
/**
 * Auction.View.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ShowProductsResponseBidDto } from './showProductsResponseBidDto';
import { ProductCategory } from './productCategory';
import { ShowProductsResponseUserDto } from './showProductsResponseUserDto';


export interface ShowProductsResponseProductDto { 
    productId?: string | null;
    productName?: string | null;
    shortDescription?: string | null;
    detailedDescription?: string | null;
    categoryType?: ProductCategory;
    startingPrice?: number;
    bidEndDate?: string;
    userId?: number;
    sellerInformation?: ShowProductsResponseUserDto;
    bids?: Array<ShowProductsResponseBidDto> | null;
}